Add files via upload
Please note remark before the function about lemi=[].
Furthermore, the levels will always begin with level 0, then 00,01 then 000,001 or 010,011 and so on. So the "depth" is the number of digits. Notice - there are cases with only one child level, in case the other child leads to a 3-phase consumer. In that case, it is not considered a cluster, and thus will not appear in the laminar set.